% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/default_forest_panels.R
\name{default_forest_panels}
\alias{default_forest_panels}
\title{Default panels for forest_model}
\usage{
default_forest_panels(model, factor_separate_line = FALSE,
  measure = "Estimate")
}
\arguments{
\item{model}{model object to guess label and determine defaults}

\item{factor_separate_line}{changes defaults for widths of variable depending on whether
  factors have their name on separate line}

\item{measure}{label for main forest plot}
}
\value{
`list` ready to be passed to `forest_model`
}
\description{
Default panels for forest_model
}

