% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_breaks.R
\name{forest_breaks}
\alias{forest_breaks}
\title{Calculate default breaks for limits}
\usage{
forest_breaks(limits, trans = I)
}
\arguments{
\item{limits}{limits of plot}

\item{trans}{transformation that will be used on the limits}
}
\value{
a vector with breaks ready to pass to \code{\link{panel_forest_plot}}
}
\description{
This function does not work as well as grDevices::axisTicks and so that should
be used instead.
}
