% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_model_format_options.R
\name{forest_model_format_options}
\alias{forest_model_format_options}
\title{Create format options for \code{forest_model}}
\usage{
forest_model_format_options(colour = "black", color = NULL,
  shape = 15, text_size = 5, banded = TRUE)
}
\arguments{
\item{colour}{colour of the point estimate and error bars}

\item{color}{alias for colour}

\item{shape}{shape of the point estimate}

\item{text_size}{text size in mm}

\item{banded}{whether to show light grey bands behind alternate rows}
}
\value{
list of format options
}
\description{
Create format options for \code{forest_model}
}
