% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_panel.R
\name{forest_panel}
\alias{forest_panel}
\title{Create definition of a panel for \code{forest_model}}
\usage{
forest_panel(
  width,
  item = c("", "forest", "vline"),
  display = NULL,
  display_na = NULL,
  hjust = NULL,
  heading = NULL,
  fontface = NULL,
  linetype = NULL,
  line_x = NULL,
  parse = NULL,
  width_group = NULL
)
}
\arguments{
\item{width}{relative width of the panel}

\item{item}{specification of which type of item to use; overridden if display is not missing}

\item{display}{bare expression that specifies the variable or expresion to display}

\item{display_na}{what to display if a value is \code{NA}}

\item{hjust}{horizontal justification}

\item{heading}{heading to be used (defaults to the variable name)}

\item{fontface}{fontface to use}

\item{linetype}{line type to use}

\item{line_x}{position for dashed line in forest plot}

\item{parse}{whether text should be parsed as expressions}

\item{width_group}{grouping used when recalcualting widths of panels}
}
\value{
panel definition as a list
}
\description{
Create definition of a panel for \code{forest_model}
}
