% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_panels.R
\name{forest_panels}
\alias{forest_panels}
\title{Generate panels for forest plots}
\usage{
forest_panels(..., margin = 0.03)
}
\arguments{
\item{...}{panels to variables in data}

\item{margin}{margin to leave at left and right edges}
}
\value{
a panels list ready for \code{\link{forest_model}} or \code{\link{forest_rma}}
}
\description{
Generate panels for forest plots
}
