% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest_rma.R
\name{get_data_for_rma}
\alias{get_data_for_rma}
\title{Extract data from individual rma model}
\usage{
get_data_for_rma(model, study_labels = NULL, model_label = NULL,
  point_size = NULL, additional_data = NULL,
  show_individual_studies = TRUE, show_stats = NULL)
}
\arguments{
\item{model}{a single \code{\link[metafor]{rma}} object or a \code{list} of them}

\item{study_labels}{a character vector of study labels or list of character vectors the same length as \code{model}}

\item{model_label}{a single model label or character vector of model labels the same length as \code{model}}

\item{point_size}{a numeric vector with the point sizes for the individual studies, or a single value used for
all studies, or a list of numeric vectors if more than one model is to be plotted}

\item{additional_data}{a \code{data.frame} of additional data that can be referenced for the data
shown in the panels of the forest plot}

\item{show_individual_studies}{whether to show the individual studies (the default) or just the summary diamond}

\item{show_stats}{a \code{list} of stats to show at the bottom of the forest plot for e.g. heterogeneity}
}
\value{
a data.frame with the extracted data
}
\description{
Extract data from individual rma model
}

