% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest_model.R
\name{panel_forest_plot}
\alias{panel_forest_plot}
\title{Plot a forest plot with panels of text}
\usage{
panel_forest_plot(forest_data, forest_mapping = aes(estimate, y, xmin =
  conf.low, xmax = conf.high), panels, trans = I, funcs = NULL,
  forest_theme = theme_ ~ forest(), colour = "black", shape = 15,
  banded = TRUE)
}
\arguments{
\item{forest_data}{\code{data.frame} with the data needed for both the plot and text}

\item{forest_mapping}{mapping aesthetic created using \code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_string}}}

\item{panels}{\code{list} with details of the panels that make up the plot (See Details)}

\item{trans}{transform for scales}

\item{funcs}{optional list of functions required for formatting \code{panels$display}}

\item{forest_theme}{theme to apply to the plot}

\item{colour}{colour of the point estimate and error bars}

\item{shape}{shape of the point estimate}

\item{banded}{whether to show light grey bands behind alternate rows}
}
\value{
A ggplot ready for display or saving
}
\description{
Plot a forest plot with panels of text
}

