% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest_model.R
\name{default_forest_panels}
\alias{default_forest_panels}
\title{Default panels for forest_model}
\usage{
default_forest_panels(measure = "Hazard ratio",
  factor_separate_line = FALSE)
}
\arguments{
\item{measure}{label for main forest plot}

\item{factor_separate_line}{changes defaults for widths of variable depending on whether
  factors have their name on separate line}
}
\value{
`list` ready to be passed to `forest_model`
}
\description{
Default panels for forest_model
}

