% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_forest_panels.R
\name{default_forest_panels}
\alias{default_forest_panels}
\title{Default panels for forest_model}
\usage{
default_forest_panels(model = NULL, factor_separate_line = FALSE,
  measure = NULL, trans_char = "I")
}
\arguments{
\item{model}{model object to guess label and determine defaults}

\item{factor_separate_line}{changes defaults for widths of variable depending on whether
factors have their name on separate line}

\item{measure}{label for main forest plot}

\item{trans_char}{character representation of transform for axes}
}
\value{
`list` ready to be passed to `forest_model`
}
\description{
Default panels for forest_model
}
