% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest_model.R
\name{forest_model}
\alias{forest_model}
\title{Produce a forest plot based on a regression model}
\usage{
forest_model(model, panels = default_forest_panels(), exponentiate = NULL,
  colour = "black", shape = 15, banded = TRUE, funcs = NULL,
  factor_separate_line = FALSE)
}
\arguments{
\item{model}{regression model produced by \code{\link[stats]{lm}},
\code{\link[stats]{glm}}, \code{\link[survival]{coxph}}}

\item{panels}{\code{data.frame} with details of the panels that make up the plot (See Details)}

\item{exponentiate}{whether the numbers on the x scale should be exponentiated for plotting}

\item{colour}{colour of the point estimate and error bars}

\item{shape}{shape of the point estimate}

\item{banded}{whether to show light grey bands behind alternate rows}

\item{funcs}{optional list of functions required for formatting \code{panels$display}}

\item{factor_separate_line}{whether to show the factor variable name on a separate line}
}
\value{
A ggplot ready for display or saving
}
\description{
Produce a forest plot based on a regression model
}
\details{
This function takes the model output from one of the common model functions in
  R (e.g. \code{\link[stats]{lm}}, \code{\link[stats]{glm}},
 \code{\link[survival]{coxph}}).

 The \code{panels} parameter is a \code{data.frame} with columns
 \code{display} and \code{width} and, optionally, \code{display_reference},
 \code{heading} and \code{hjust}.
 \code{display} indicates which column to display as text from the standard ones produced by
 \code{\link[broom]{tidy}} and in addition
 \code{variable} (the term in the model, for factors without the level),
 \code{level} (the level of factors),
 \code{reference} (TRUE for the reference level of a factor).
 It can also be a formula using these columns. The function \code{trans} is definded to be the
 transformation between the coefficients and the scales (e.g. \code{exp}). Other functions not
 in base R can be provided as a \code{list} with the parameter \code{funcs}.
 \code{display_reference} allows for an alternative display for reference levels of factors
 which typically do not have data within the model output.
}
\examples{
library("survival")
library("dplyr")
pretty_lung <- lung \%>\%
  transmute(time,
            status,
            Age = age,
            Sex = factor(sex, labels = c("Male", "Female")),
            ECOG = factor(lung$ph.ecog),
            `Meal Cal` = meal.cal)

print(forest_model(coxph(Surv(time, status) ~ ., pretty_lung)))

data_for_lm <- data_frame(x = rnorm(100, 4),
                          y = rnorm(100, 3, 0.5),
                          z = rnorm(100, 2, 2),
                          outcome = 3 * x - 2 * y + 4 * z + rnorm(100, 0, 0.1))

print(forest_model(lm(outcome ~ ., data_for_lm), default_forest_panels("Estimate")))

data_for_logistic <- data_for_lm \%>\% mutate(
  outcome = (0.5 * (x - 4) * (y - 3) * (z - 2) + rnorm(100, 0, 0.05)) > 0.5
)

print(forest_model(glm(outcome ~ ., binomial(), data_for_logistic),
  default_forest_panels("Odds ratio")))
}

