% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel_forest_plot.R
\name{panel_forest_plot}
\alias{panel_forest_plot}
\title{Plot a forest plot with panels of text}
\usage{
panel_forest_plot(forest_data, mapping = aes(estimate, xmin = conf.low, xmax =
  conf.high), panels = default_forest_panels(), trans = I, funcs = NULL,
  forest_theme = theme_forest(), colour = "black", shape = 15,
  banded = TRUE, limits = NULL, breaks = NULL)
}
\arguments{
\item{forest_data}{\code{data.frame} with the data needed for both the plot and text}

\item{mapping}{mapping aesthetic created using \code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_string}}}

\item{panels}{\code{list} with details of the panels that make up the plot (See Details)}

\item{trans}{transform for scales}

\item{funcs}{optional list of functions required for formatting \code{panels$display}}

\item{forest_theme}{theme to apply to the plot}

\item{colour}{colour of the point estimate and error bars}

\item{shape}{shape of the point estimate}

\item{banded}{whether to show light grey bands behind alternate rows}

\item{limits}{limits of the forest plot on the X-axis (taken as the range of the data
by default)}

\item{breaks}{breaks to appear on the X-axis (note these will be exponentiated
if \code{exponentiate == TRUE})}
}
\value{
A ggplot ready for display or saving
}
\description{
Plot a forest plot with panels of text
}

